requirements.txt
# 数字人交互系统依赖

# Web框架
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
python-multipart==0.0.6

# HTTP客户端
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# AI/ML
openai==1.10.0
ollama==0.1.6

# 语音合成
edge-tts==6.1.9

# 语音识别 (可选 - 根据需要安装)
# OpenAI Whisper (本地运行，免费)
openai-whisper==20231117
# 或使用OpenAI API (需要API Key)
# 已包含在 openai 包中

# 音频处理
pydub==0.25.1
soundfile==0.12.1

# 图像/视频处理
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 工具
loguru==0.7.2
tenacity==8.2.3
aiofiles==23.2.1

# 可选：本地LLM支持
# torch==2.1.2
# transformers==4.37.0